Refactor Codex account coordination
