+++ rust/src/codex_accounts/mod.rs
@@ -14,10 +14,13 @@
 pub mod account_manager;
 pub mod api;
 pub mod codex_desktop;
+pub(crate) mod credentials;
+mod fetch_coordination;
 pub mod file_locations;
 pub mod login_runner;
 pub mod models;
 pub mod stores;
+mod switch_runtime;
 
 // Refreshes may rotate auth.json. Keep account replacement exclusive with
 // those reads/writes while allowing different account lanes to fetch together.
@@ -25,14 +28,16 @@ pub(crate) static CREDENTIAL_OPERATIONS: tokio::sync::RwLock<()> =
     tokio::sync::RwLock::const_new(());
 
 pub use account_manager::{CodexAccountManager, CodexAccountManagerError, CodexSwitchResult};
-pub use api::{AuthBackedIdentity, AuthCredentials, CodexAccountApi, CodexApiError, load_identity};
+pub use api::{CodexAccountApi, CodexApiError};
 pub use codex_desktop::{
     CodexDesktopControlError, build_restart_command, build_restart_script,
     encode_powershell_script, restart_codex_desktop,
 };
+pub use credentials::{AuthBackedIdentity, AuthCredentials, load_identity};
 pub use login_runner::{CodexLoginOutcome, CodexLoginResult, ManagedLoginProcess};
 pub use models::{
     AccountUsageSnapshot, CodexAccount, CodexAccountSource, CreditsBalanceSnapshot,
     RemovedAccountIdentity, UsageWindowSnapshot, utc_now,
 };
 pub use stores::{AccountStore, SnapshotStore};
+pub use switch_runtime::CodexAccountRuntime;
